% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dflatExample}
\alias{dflatExample}
\title{A subset of the DFLAT gene sets}
\format{A list of lists of gene identifiers, indexed by gene set name}
\source{
\url{http://bcb.cs.tufts.edu/dflat//}
}
\usage{
dflatExample
}
\description{
20 selected gene sets from the DFLAT biological process gene sets
}
\keyword{datasets}
