% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gse32472Example}
\alias{gse32472Example}
\title{A subset of the GSE32472 BPD data set}
\format{A list
\describe{
  \item{data}{a matrix of 97 samples and 654 genes}
  \item{age}{a list of sample gestational ages, indexed by sample id}
  \item{bpd}{a list of statuses, indexed by sample id}
  \item{ctrl}{a list of sample ids without BPD}
  \item{test}{a list of sample ids with BPD}
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE32472}
}
\usage{
gse32472Example
}
\description{
A dataset of 97 normal or BPD samples at 5 days of life, with gestational age, class, and gene expression for 654 genes
}
\keyword{datasets}
