% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gse32472ExampleTempoResults}
\alias{gse32472ExampleTempoResults}
\title{TEMPO results for the GSE32472 subset example data set}
\format{A list
\describe{
  \item{ctrl}{a list of control sample ids used for scoring, identical to gse32472Example$ctrl}
  \item{test}{a list of test sample ids used for scoring, identical to gse32472Example$test}
  \item{train}{a list of normal sample ids used for model training, identical to gse32472Example$ctrl}
  \item{scores}{a data frame of scores and signicance values for each gene set}
  \item{pred}{a matrix of age predictions for each sample and gene set}
  \item{Y}{age for each sample, identical to gse32472Example$age}
  \item{genesets}{the gene sets used in the analysis, identical to dflatExample}
  \item{reported}{a list of the gene sets which meet the significance thresholds for reporting}
}}
\usage{
gse32472ExampleTempoResults
}
\description{
The results data structure generated by tempo.run with default settings on the gse32472Example data set
}
\keyword{datasets}
