% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempoOutput.R
\name{tempo.mkplot}
\alias{tempo.mkplot}
\title{Make a plot for a specified gene sets}
\usage{
tempo.mkplot(results, geneset)
}
\arguments{
\item{results}{The results object returned by \code{\link{tempo.runInstance}}.or \code{\link{tempo.run}}}

\item{geneset}{The name of a gene set}
}
\description{
Generate a plot of actual vs. predicted age, broken down by control sample or test sample, for a single specified gene set
}
\examples{
data("gse32472ExampleTempoResults")
tempo.mkplot(gse32472ExampleTempoResults, "cochlea development")
}
