% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempoOutput.R
\name{tempo.writeOutput}
\alias{tempo.writeOutput}
\title{Write output}
\usage{
tempo.writeOutput(results, filename)
}
\arguments{
\item{results}{The results object returned by \code{\link{tempo.runInstance}}.}

\item{filename}{The prefix for the filename.  Writes filename.pdf and filename.table (a tab-delimited text file) for all reported gene sets}
}
\description{
Write the score data frame and plots of actual vs. predicted age for all reported gene sets to disk.
}
\examples{
# In this example, text output will be written to tempfile.table and
# plots printed to tempfile.pdf
data("gse32472ExampleTempoResults")
tempo.writeOutput(gse32472ExampleTempoResults, tempfile(tmpdir = tempdir()))
}
